requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
CacheControl==0.14.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.0
firebase_admin==7.1.0
gcloud==0.18.3
google-api-core==2.28.1
google-auth==2.47.0
google-cloud-core==2.5.0
google-cloud-firestore==2.22.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jwcrypto==1.5.6
msgpack==1.1.2
oauth2client==4.1.3
proto-plus==1.27.0
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.3.1
Pyrebase4==4.8.0
python-dotenv==1.2.1
python-jwt==4.1.0
requests==2.32.5
requests-toolbelt==0.10.1
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==1.26.20
uvicorn==0.40.0

# Socket.io for real-time chatbot
python-socketio[asyncio]==5.11.0
aiohttp==3.9.5

# OpenRouter/OpenAI client for LLM
openai>=1.50.0

# News RSS feed parsing
feedparser==6.0.11
beautifulsoup4==4.12.3

# Background task scheduling
apscheduler==3.10.4

# Image storage
cloudinary>=1.44.0

pytest>=8.0
pytest-asyncio>=0.24